[BE]Refactor: 성공 및 에러 응답 유틸리티 함수 분리
