+++ backend/src/middlewares/errorHandler.ts
@@ -1,26 +1,27 @@
 import { Request, Response, NextFunction } from 'express';
+import { createErrorResponse } from '../utils/responseUtils.js';
 
 interface AppError extends Error {
   statusCode?: number;
 }
 
-const errorHandler = (
-  err: AppError, 
-  req: Request, 
-  res: Response, 
-  _next: NextFunction
-): void => {
+const errorHandler = (err: AppError, req: Request, res: Response, _next: NextFunction): void => {
   console.error('Error:', err.message);
   console.error('Stack:', err.stack);
 
   const statusCode = err.statusCode || 500;
   const message = err.message || '서버 내부 오류가 발생했습니다.';
 
-  res.status(statusCode).json({
-    status: 'error',
-    message,
-    ...(process.env.NODE_ENV === 'development' && { stack: err.stack }),
-  });
+  if (process.env.NODE_ENV === 'development') {
+    res.status(statusCode).json({
+      status: 'error',
+      message,
+      stack: err.stack,
+    });
+    return;
+  }
+
+  createErrorResponse(res, statusCode, message);
 };
 
-export default errorHandler; 
+export default errorHandler;
